# doc/openapi.yaml
openapi: "3.0.2"
info:
  title: Friend Finder
  version: '1.0'
  description: "L'API de FriendFinder est une API REST. Elle a des URL orientées ressources, elle renvoie des données codées en JSON, elle utilise les codes de réponses standards HTTP, et une authentification."

  #termsOfService: "https://openweathermap.org/terms"
  license:
    name: "Repo Git"
    url: "https://github.com/EnseirbTelecom/pg219-2020-dark-heroes/tree/master/doc"
servers:
  - url: https://localhost:3000
security:
  - friendFinder_auth: []
...
components:
  securitySchemes:
    friendFinder_auth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: "L'utilisateur doit s'identifier pour accéder aux requêtes."
  schemas:
    Hist_position:
      type: object
      properties:
        hist_lat:
          type: string
        hist_long:
          type: string
        hist_date:
          type: string
          format: date
        message:
          type: string
...
    User:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        first_name:
          type: string
        last_name:
          type: string
        pseudo:
          type: string
        birthday:
          type: string
          format: date
        gender:
          type: string
        current_lat:
          type: string
        current_long:
          type: string
        activation_date:
          type: string
          format: date
        duration:
          type: number
          minimum: 0
        hist_positions:
          $ref: "#/components/schemas/Hist_position"
        friends:
          type: array
          items:
            type: string
        friend_request:
          type: array
          items:
            type: string
        friend_pending:
          type: array
          items:
            type: string
paths:
  /profile:
    get:
      tags:
        - Utilisateur
      summary: "Accéder à son profil"
      operationId: "getMyProfile"
      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  user:
                    $ref: "#/components/schemas/User"
        "404":
          description: "Utilisateur introuvable"
          content:
            json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "User not found"
                  status:
                    type: integer
                    example: 404
        "403":
          description: "Erreur d'autorisation"
          content:
            json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "token not valid"
                  status:
                    type: integer
                    example: 403
  /delProfile:
    delete:
      tags:
        - Utilisateur
      operationId: "deleteMyProfile"
      summary: "Supprimer son profil"
      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  state:
                    type: string
                    example: "user deleted"


        "403":
          description: "Erreur d'autorisation"
          content:
            json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "token not valid"
                  status:
                    type: integer
                    example: 403
  /settings:
    post:
      tags:
        - Utilisateur
      summary: "Changer de mot de passe"
      operationId: "changepassword"
      requestBody:
        description: "JSON contenant l'ancien et le nouveau mot de passe"
        required: true
        content:
          json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                  format: password
                new_password:
                  type: string
                  format: password

      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
        "403":
          description: "Erreur d'autorisation ou ancien mot de passe incorrect"
          content:
            json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "token not valid"
                  status:
                    type: integer
                    example: 403
  /Users:
    post:
      tags:
        - Utilisateur
      summary: "Ajouter un utilisateur"
      description: ""
      operationId: "addUser"
      requestBody:
        description: "Structure de User de l'utilisateur à ajouter"
        required: true
        content:
          json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  state:
                    type: string
                    example: "Your account has been created."
                  status:
                    type: integer
                    example: 200

        "400":
          description: "Informations manquantes, ou pseudo/email déjà attribué."
          content:
            json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Please, fill in the form proprely."
                  status:
                    type: integer
                    example: 400
...
  /Connect:
    post:
      tags:
        - Connexion
      summary: "Se connecter"
      operationId: "connectUser"
      requestBody:
        description: "Structure User de l'utilisateur"
        required: true
        content:
          json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  state:
                    type: string
                    example: "connected"
                  token:
                    type: number
                    example: ""
                  status:
                    type: integer
                    example: 200
        "403":
          description: "Erreur de connexion"
          content:
            json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "failed to connect"
        "401":
          description: "Utilisateur introuvable"
          content:
            json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "user unknown"
  /isConnect:
    post:
      tags:
        - Connexion
      summary: "Vérifier l'état de connexion"
      operationId: "isConnect"
      responses:
        "200":
          description: "Token valide"
          content:
            json:
              schema:
                type: object
                properties:
                  state:
                    type: string
                    example: "connected, token valid"
        "403":
          description: "Erreur de connexion"
          content:
            json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "token not valid"